+++ agent/php5/agent/openrasp_agent_manager.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "utils/JsonReader.h"
+#include "utils/json_reader.h"
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/wait.h>
