+++ agent/php5/agent/shared_config_block.h
@@ -18,7 +18,7 @@
 
 #include "openrasp.h"
 #include "openrasp_hook.h"
-#include "utils/DoubleArrayTrie.h"
+#include "utils/double_array_trie.h"
 #include <string>
 
 namespace openrasp
