+++ agent/php5/agent/shared_config_manager.h
@@ -21,9 +21,9 @@
 #include "base_manager.h"
 #include <memory>
 #include <map>
-#include "utils/ReadWriteLock.h"
+#include "utils/read_write_lock.h"
 #include "shared_config_block.h"
-#include "utils/BaseReader.h"
+#include "utils/base_reader.h"
 
 namespace openrasp
 {
