+++ agent/php5/utils/base_reader.h
