+++ agent/php5/utils/double_array_trie.h
