+++ agent/php7/config.m4
@@ -508,8 +508,8 @@ int main() {
     openrasp_v8_utils.cc \
     openrasp_security_policy.cc \
     openrasp_ini.cc \
-    utils/ReadWriteLock.cc \
-    utils/DoubleArrayTrie.cc \
+    utils/read_write_lock.cc \
+    utils/double_array_trie.cc \
     utils/string.cc \
     utils/digest.cc \
     utils/regex.cc \
@@ -517,8 +517,8 @@ int main() {
     utils/file.cc \    
     utils/time.cc \
     utils/net.cc \
-    utils/JsonReader.cc \
-    utils/YamlReader.cc \
+    utils/json_reader.cc \
+    utils/yaml_reader.cc \
     utils/utf.cc \
     utils/hostname.cc \
     agent/base_manager.cc \
